fix: keep MAP20001 shortcut tests under app test file-size ceiling (#4479)
Split demo overlay shells out of the shared harness and route shortcut
callback pins through provinceDetailCallbacks with optional topology so
repo.app_test_file_size stays ≤400.
