+++ app/test/province_overlay_test_harness.dart
@@ -6,8 +6,6 @@ import 'package:colonizethis_logic/colonizethis_logic.dart'
     show
         PlayerView,
         ProvinceImprovableCommodityCount,
-        ProvinceTileCapitalLinkPreview,
-        VisibilityLevel,
         buildPlayerView;
 import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
@@ -15,12 +13,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_fixtures/demo/province_overlay_demo_data.dart'
-    show
-        demoGameForOverlay,
-        demoHumanPlayerViewForOverlay,
-        demoRegionForOverlay,
-        sampleProvinceIdForOverlay,
-        sampleTileKeyForProvinceOverlay;
+    show demoHumanPlayerViewForOverlay, demoRegionForOverlay;
 import 'package:colonizethis_app_l10n/l10n/l10n.dart';
 import 'package:colonizethis_app/features/game/flame/overlays/province_detail_overlay_host_support_tile_connectivity.dart'
     show ProvinceTileConnectivityDisplay;
@@ -281,145 +274,4 @@ Future<void> pumpProvinceOverlayAtDarkTheme(
   );
   await tester.pump();
   await tester.pump(const Duration(milliseconds: 16));
-}
-
-/// Overlay with the demo fixture player view and a revealed sample land tile.
-Widget buildProvinceOverlayWithRevealedDemoTile({int roadLevel = 0}) {
-  final base = demoGameForOverlay;
-  final tileKey = sampleTileKeyForProvinceOverlay;
-  final game = roadLevel == 0
-      ? base
-      : gameWithRoadLevelOnTile(
-          base: base,
-          tileKey: tileKey,
-          roadLevel: roadLevel,
-        );
-  return buildProvinceOverlayDarkThemeShell(
-    game: game,
-    displayId: provinceIdFromTileKey(tileKey),
-    selectedTileKey: tileKey,
-    playerView: demoOverlayPlayerView(base),
-  );
-}
-
-/// Overlay with demo fixture, full player view, and configurable road level.
-Widget buildProvinceOverlayWithRoadLevelFullPlayerView({
-  required int roadLevel,
-}) {
-  final base = demoGameForOverlay;
-  final tileKey = sampleTileKeyForProvinceOverlay;
-  final game = gameWithRoadLevelOnTile(
-    base: base,
-    tileKey: tileKey,
-    roadLevel: roadLevel,
-  );
-  return buildProvinceOverlayDarkThemeShell(
-    game: game,
-    displayId: provinceIdFromTileKey(tileKey),
-    selectedTileKey: tileKey,
-    playerView: demoOverlayPlayerView(base),
-  );
-}
-
-/// Overlay with demo fixture, demo player view, and optional inline action icons.
-Widget buildProvinceOverlayWithRoadLevelDemoFixture({
-  required int roadLevel,
-  bool showExploreActionIcon = false,
-  bool exploreActionEnabled = false,
-  bool showProspectActionIcon = false,
-  bool prospectActionEnabled = false,
-  bool showBuildImprovementActionIcon = false,
-  bool buildImprovementActionEnabled = false,
-}) {
-  final base = demoGameForOverlay;
-  final tileKey = sampleTileKeyForProvinceOverlay;
-  final game = gameWithRoadLevelOnTile(
-    base: base,
-    tileKey: tileKey,
-    roadLevel: roadLevel,
-  );
-  return buildProvinceOverlayDarkThemeShell(
-    game: game,
-    displayId: sampleProvinceIdForOverlay,
-    selectedTileKey: tileKey,
-    showExploreActionIcon: showExploreActionIcon,
-    exploreActionEnabled: exploreActionEnabled,
-    onExploreWithExplorerTap: () {},
-    showProspectActionIcon: showProspectActionIcon,
-    prospectActionEnabled: prospectActionEnabled,
-    onProspectWithExplorerTap: () {},
-    showBuildImprovementActionIcon: showBuildImprovementActionIcon,
-    buildImprovementActionEnabled: buildImprovementActionEnabled,
-    onBuildImprovementTap: () {},
-  );
-}
-
-/// Overlay with a province-context display id and a sea-cell selected tile.
-/// Returns `null` when the demo fixture cannot supply the port-harbor case.
-Widget? buildProvinceOverlayWithSeaCellAtLandProvince() {
-  final base = demoGameForOverlay;
-  final region = demoRegionForOverlay;
-  final humanPlayerId = base.players.first.id;
-  final playerView = demoOverlayPlayerView(base);
-
-  String? landDisplayId;
-  for (final cell in region.cells) {
-    if (!cell.isSea) {
-      landDisplayId = '${region.regionId}|${cell.regionCellId}';
-      break;
-    }
-  }
-
-  String? seaTileKey;
-  for (final cell in region.cells) {
-    if (!cell.isSea) continue;
-    if (cell.visibility == TileVisibility.unrevealed) continue;
-    final candidate =
-        '${region.regionId}|${cell.regionCellId}|${cell.x}|${cell.y}';
-    if (playerView.visibilityForTile(candidate) != VisibilityLevel.unknown) {
-      seaTileKey = candidate;
-      break;
-    }
-  }
-  if (landDisplayId == null || seaTileKey == null) return null;
-
-  return buildProvinceOverlayDarkThemeShell(
-    game: base,
-    displayId: landDisplayId,
-    selectedTileKey: seaTileKey,
-    playerView: playerView,
-  );
-}
-
-/// Searches the demo region for a revealed land tile with no resource.
-({String displayId, String tileKey})? findRevealedLandTileWithoutResource() {
-  final game = demoGameForOverlay;
-  final region = demoRegionForOverlay;
-  final playerView = demoOverlayPlayerView(game);
-  final resourceByTile = game.worldState.resourceByTileKey;
-  for (final cell in region.cells) {
-    if (cell.isSea) continue;
-    if (cell.resourceId != null) continue;
-    final tk = '${region.regionId}|${cell.regionCellId}|${cell.x}|${cell.y}';
-    if (resourceByTile[tk] != null) continue;
-    if (playerView.visibilityForTile(tk) != VisibilityLevel.fullyVisible) {
-      continue;
-    }
-    final displayId = '${region.regionId}|${cell.regionCellId}';
-    return (displayId: displayId, tileKey: tk);
-  }
-  return null;
-}
-
-/// Overlay with a revealed land tile that has no visible resource.
-Widget? buildProvinceOverlayWithRevealedNoResourceTile() {
-  final pick = findRevealedLandTileWithoutResource();
-  if (pick == null) return null;
-  final game = demoGameForOverlay;
-  return buildProvinceOverlayDarkThemeShell(
-    game: game,
-    displayId: pick.displayId,
-    selectedTileKey: pick.tileKey,
-    playerView: demoOverlayPlayerView(game),
-  );
 }
