+++ app/test/province_overlay_tile_resource_row_dark_token_test.dart
@@ -30,6 +30,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 
 import 'province_overlay_test_harness.dart';
+import 'province_overlay_test_harness_demo_shells.dart';
 
 /// Returns the `Text` widget whose `data` exactly equals [text]. Used
 /// for the literal Resource row children (`Resource: ` prefix and the
