+++ app/test/province_overlay_tile_section_body_rows_dark_tokens_test.dart
@@ -27,6 +27,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
 
 import 'province_overlay_test_harness.dart';
+import 'province_overlay_test_harness_demo_shells.dart';
 
 /// Returns the `Text` widget whose `data` starts with [prefix] (e.g.
 /// `Coordinates: `, `Terrain: `, `Civilian units (province): `). The Tile
